Relocate ServiceTypesQuery tests to tests/services/test_types
