+++ tests/test_init.py
@@ -6,20 +6,18 @@
 
 import errno
 import logging
-import os
 import socket
 import time
 import unittest
 import unittest.mock
-from threading import Event
 from typing import Dict, Optional  # noqa # used in type hints
 
 import pytest
 
 import zeroconf as r
-from zeroconf import ServiceBrowser, ServiceInfo, Zeroconf, ZeroconfServiceTypes, const
+from zeroconf import ServiceBrowser, ServiceInfo, Zeroconf, const
 
-from . import has_working_ipv6, _clear_cache, _inject_response
+from . import _clear_cache
 
 log = logging.getLogger('zeroconf')
 original_logging_level = logging.NOTSET
@@ -457,135 +455,6 @@ class TestServiceRegistry(unittest.TestCase):
         assert registry.get_infos_type(type_) == [info]
         assert registry.get_infos_server("ash-2.local.") == [info]
         assert registry.get_types() == [type_]
-
-
-class ServiceTypesQuery(unittest.TestCase):
-    def test_integration_with_listener(self):
-
-        type_ = "_test-srvc-type._tcp.local."
-        name = "xxxyyy"
-        registration_name = "%s.%s" % (name, type_)
-
-        zeroconf_registrar = Zeroconf(interfaces=['127.0.0.1'])
-        desc = {'path': '/~paulsm/'}
-        info = ServiceInfo(
-            type_,
-            registration_name,
-            80,
-            0,
-            0,
-            desc,
-            "ash-2.local.",
-            addresses=[socket.inet_aton("10.0.1.2")],
-        )
-        zeroconf_registrar.register_service(info)
-
-        try:
-            service_types = ZeroconfServiceTypes.find(interfaces=['127.0.0.1'], timeout=0.5)
-            assert type_ in service_types
-            _clear_cache(zeroconf_registrar)
-            service_types = ZeroconfServiceTypes.find(zc=zeroconf_registrar, timeout=0.5)
-            assert type_ in service_types
-
-        finally:
-            zeroconf_registrar.close()
-
-    @unittest.skipIf(not has_working_ipv6(), 'Requires IPv6')
-    @unittest.skipIf(os.environ.get('SKIP_IPV6'), 'IPv6 tests disabled')
-    def test_integration_with_listener_v6_records(self):
-
-        type_ = "_test-srvc-type._tcp.local."
-        name = "xxxyyy"
-        registration_name = "%s.%s" % (name, type_)
-        addr = "2606:2800:220:1:248:1893:25c8:1946"  # example.com
-
-        zeroconf_registrar = Zeroconf(interfaces=['127.0.0.1'])
-        desc = {'path': '/~paulsm/'}
-        info = ServiceInfo(
-            type_,
-            registration_name,
-            80,
-            0,
-            0,
-            desc,
-            "ash-2.local.",
-            addresses=[socket.inet_pton(socket.AF_INET6, addr)],
-        )
-        zeroconf_registrar.register_service(info)
-
-        try:
-            service_types = ZeroconfServiceTypes.find(interfaces=['127.0.0.1'], timeout=0.5)
-            assert type_ in service_types
-            _clear_cache(zeroconf_registrar)
-            service_types = ZeroconfServiceTypes.find(zc=zeroconf_registrar, timeout=0.5)
-            assert type_ in service_types
-
-        finally:
-            zeroconf_registrar.close()
-
-    @unittest.skipIf(not has_working_ipv6(), 'Requires IPv6')
-    @unittest.skipIf(os.environ.get('SKIP_IPV6'), 'IPv6 tests disabled')
-    def test_integration_with_listener_ipv6(self):
-
-        type_ = "_test-srvc-type._tcp.local."
-        name = "xxxyyy"
-        registration_name = "%s.%s" % (name, type_)
-
-        zeroconf_registrar = Zeroconf(ip_version=r.IPVersion.V6Only)
-        desc = {'path': '/~paulsm/'}
-        info = ServiceInfo(
-            type_,
-            registration_name,
-            80,
-            0,
-            0,
-            desc,
-            "ash-2.local.",
-            addresses=[socket.inet_aton("10.0.1.2")],
-        )
-        zeroconf_registrar.register_service(info)
-
-        try:
-            service_types = ZeroconfServiceTypes.find(ip_version=r.IPVersion.V6Only, timeout=0.5)
-            assert type_ in service_types
-            _clear_cache(zeroconf_registrar)
-            service_types = ZeroconfServiceTypes.find(zc=zeroconf_registrar, timeout=0.5)
-            assert type_ in service_types
-
-        finally:
-            zeroconf_registrar.close()
-
-    def test_integration_with_subtype_and_listener(self):
-        subtype_ = "_subtype._sub"
-        type_ = "_type._tcp.local."
-        name = "xxxyyy"
-        # Note: discovery returns only DNS-SD type not subtype
-        discovery_type = "%s.%s" % (subtype_, type_)
-        registration_name = "%s.%s" % (name, type_)
-
-        zeroconf_registrar = Zeroconf(interfaces=['127.0.0.1'])
-        desc = {'path': '/~paulsm/'}
-        info = ServiceInfo(
-            discovery_type,
-            registration_name,
-            80,
-            0,
-            0,
-            desc,
-            "ash-2.local.",
-            addresses=[socket.inet_aton("10.0.1.2")],
-        )
-        zeroconf_registrar.register_service(info)
-
-        try:
-            service_types = ZeroconfServiceTypes.find(interfaces=['127.0.0.1'], timeout=0.5)
-            assert discovery_type in service_types
-            _clear_cache(zeroconf_registrar)
-            service_types = ZeroconfServiceTypes.find(zc=zeroconf_registrar, timeout=0.5)
-            assert discovery_type in service_types
-
-        finally:
-            zeroconf_registrar.close()
 
 
 def test_ptr_optimization():
